Split lexer into token and expansion modules
